+++ erl/genserver/gen_server.go
@@ -391,11 +391,3 @@ func (gs *GenServerS[STATE]) handleCastRequest(self erl.PID, msg CastRequest) (e
 	}
 	return nil
 }
-
-
-
-
-
-
-
-